+++ src/components/util/DotSeparator.svelte
@@ -1,0 +1,1 @@
+<div class="w-1 h-1 rounded-full bg-dark-secondary/50 dark:bg-secondary/50"></div>