+++ src/store/store.js
@@ -8,6 +8,7 @@ export const updateRate = writable(15000);
 export const entryStore = writable(10);
 export const sortDirStore = writable("ascending");
 export const sortByStore = writable("rank");
+export const sortNewsBy = writable("Twitter");
 
 //Pagination
 export const pageStore = writable(1);