+++ src/util/api/fetch/fetchNewsDataById.ts
@@ -1,0 +1,17 @@
+import { getNewsDataById } from "../api";
+
+export async function fetchNewsDataById(id: string | undefined) {
+  try {
+    const response = await getNewsDataById(id);
+    if (response) {
+      return {
+        imgUrl: response.imgUrl || "-",
+      };
+    } else {
+      return { imgUrl: "-" };
+    }
+  } catch (error) {
+    console.error("Error fetching data:", error);
+    return { imgUrl: "-" };
+  }
+}