+++ src/util/api/fetch/index.ts
@@ -4,3 +4,4 @@ export * from "./fetchFiatData";
 export * from "./fetchHistoricalData";
 export * from "./fetchMarketData";
 export * from "./fetchNewsData";
+export * from "./fetchNewsDataById";