+++ src/util/utils.ts
@@ -299,3 +299,21 @@ export async function checkDataReadiness(dataArrays: any[], store: any): Promise
     }, 0);
   });
 }
+
+export function timeAgo(timestamp: string) {
+  const now: any = new Date();
+  const date: any = new Date(timestamp);
+  const seconds = Math.floor((now - date) / 1000);
+
+  const minutes = Math.floor(seconds / 60);
+  const hours = Math.floor(minutes / 60);
+  const days = Math.floor(hours / 24);
+
+  if (days > 0) {
+    return `${days} ${days === 1 ? "day" : "days"} ago`;
+  } else if (hours > 0) {
+    return `${hours} ${hours === 1 ? "hour" : "hours"} ago`;
+  } else {
+    return `${minutes} ${minutes === 1 ? "minute" : "minutes"} ago`;
+  }
+}